rename compute-sku-recommender to compute-recommender
