+++ doc/sphinx/azhelpgen/doc_source_map.json
@@ -19,7 +19,7 @@
     "cdn": "src/azure-cli/azure/cli/command_modules/cdn/_help.py",
     "cloud": "src/azure-cli/azure/cli/command_modules/cloud/_help.py",
     "cognitiveservices": "src/azure-cli/azure/cli/command_modules/cognitiveservices/_help.py",
-    "compute-sku-recommender": "src/azure-cli/azure/cli/command_modules/compute_sku_recommender/_help.py",
+    "compute-recommender": "src/azure-cli/azure/cli/command_modules/compute_recommender/_help.py",
     "config": "src/azure-cli/azure/cli/command_modules/config/_help.py",
     "configure": "src/azure-cli/azure/cli/command_modules/configure/_help.py",
     "consumption": "src/azure-cli/azure/cli/command_modules/consumption/_help.py",
