+++ src/azure-cli/azure/cli/command_modules/compute_recommender/__init__.py
@@ -6,22 +6,22 @@
 # --------------------------------------------------------------------------------------------
 
 from azure.cli.core import AzCommandsLoader
-from azure.cli.command_modules.compute_sku_recommender._help import helps  # pylint: disable=unused-import
+from azure.cli.command_modules.compute_recommender._help import helps  # pylint: disable=unused-import
 # from azure.cli.core.profiles import ResourceType  # required when using python sdk
 
 
-class ComputeSkuRecommenderCommandsLoader(AzCommandsLoader):
+class ComputeRecommenderCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
         custom_command_type = CliCommandType(
-            operations_tmpl='azure.cli.command_modules.compute_sku_recommender.custom#{}')
+            operations_tmpl='azure.cli.command_modules.compute_recommender.custom#{}')
         super().__init__(cli_ctx=cli_ctx,
                          # resource_type=ResourceType.XXX  # required when using python sdk
                          custom_command_type=custom_command_type)
 
     def load_command_table(self, args):
-        from azure.cli.command_modules.compute_sku_recommender.commands import load_command_table
+        from azure.cli.command_modules.compute_recommender.commands import load_command_table
         from azure.cli.core.aaz import load_aaz_command_table
         try:
             from . import aaz
@@ -37,8 +37,8 @@ class ComputeSkuRecommenderCommandsLoader(AzCommandsLoader):
         return self.command_table
 
     def load_arguments(self, command):
-        from azure.cli.command_modules.compute_sku_recommender._params import load_arguments
+        from azure.cli.command_modules.compute_recommender._params import load_arguments
         load_arguments(self, command)
 
 
-COMMAND_LOADER_CLS = ComputeSkuRecommenderCommandsLoader
+COMMAND_LOADER_CLS = ComputeRecommenderCommandsLoader
