+++ src/azure-cli/azure/cli/command_modules/compute_recommender/aaz/latest/compute_recommender/__init__.py
