+++ src/azure-cli/azure/cli/command_modules/compute_recommender/aaz/latest/compute_recommender/_spot_placement_recommender.py
@@ -12,13 +12,13 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "compute-sku-recommender spot-placement-recommender",
+    "compute-recommender spot-placement-recommender",
 )
 class SpotPlacementRecommender(AAZCommand):
     """Generates placement scores for Spot VM skus.
 
     :example: generate spot vm placement score example
-        az compute-sku-recommender spot-placement-recommender -l eastus --subscription ffffffff-ffff-ffff-ffff-ffffffffffff --availability-zones true --desired-locations '["eastus", "eastus2"]' --desired-count 1 --desired-sizes '[{"sku": "Standard_D2_v2"}]'
+        az compute-recommender spot-placement-recommender -l eastus --subscription ffffffff-ffff-ffff-ffff-ffffffffffff --availability-zones true --desired-locations '["eastus", "eastus2"]' --desired-count 1 --desired-sizes '[{"sku": "Standard_D2_v2"}]'
     """
 
     _aaz_info = {
