+++ src/azure-cli/azure/cli/command_modules/compute_recommender/tests/__init__.py
