+++ src/azure-cli/azure/cli/command_modules/compute_recommender/tests/latest/__init__.py
