+++ src/azure-cli/azure/cli/command_modules/compute_recommender/tests/latest/recordings/test_spot_placement_recommender_generate.yaml
@@ -8,7 +8,7 @@ interactions:
       Accept-Encoding:
       - gzip, deflate
       CommandName:
-      - compute-sku-recommender spot-placement-recommender
+      - compute-recommender spot-placement-recommender
       Connection:
       - keep-alive
       Content-Length:
@@ -24,18 +24,18 @@ interactions:
     uri: https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Compute/locations/eastus/diagnostics/spotPlacementRecommender/generate?api-version=2024-03-01-preview
   response:
     body:
-      string: '{"desiredLocations":["eastus","eastus2"],"desiredSizes":[{"sku":"Standard_D2_v2"}],"desiredCount":1,"availabilityZones":true,"placementScores":[{"sku":"Standard_D2_v2","region":"eastus","availabilityZone":"1","score":"DataNotFoundOrStale","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus","availabilityZone":"2","score":"High","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus","availabilityZone":"3","score":"High","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus2","availabilityZone":"1","score":"High","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus2","availabilityZone":"2","score":"High","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus2","availabilityZone":"3","score":"High","isQuotaAvailable":true}]}'
+      string: '{"desiredLocations":["eastus","eastus2"],"desiredSizes":[{"sku":"Standard_D2_v2"}],"desiredCount":1,"availabilityZones":true,"placementScores":[{"sku":"Standard_D2_v2","region":"eastus","availabilityZone":"1","score":"High","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus","availabilityZone":"2","score":"DataNotFoundOrStale","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus","availabilityZone":"3","score":"High","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus2","availabilityZone":"1","score":"High","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus2","availabilityZone":"2","score":"DataNotFoundOrStale","isQuotaAvailable":true},{"sku":"Standard_D2_v2","region":"eastus2","availabilityZone":"3","score":"High","isQuotaAvailable":true}]}'
     headers:
       api-supported-versions:
       - 2021-06-01-preview, 2024-03-01-preview
       cache-control:
       - no-cache
       content-length:
-      - '793'
+      - '808'
       content-type:
       - application/json; charset=utf-8
       date:
-      - Fri, 19 Apr 2024 21:23:07 GMT
+      - Tue, 23 Apr 2024 18:17:51 GMT
       expires:
       - '-1'
       pragma:
@@ -49,7 +49,7 @@ interactions:
       x-ms-ratelimit-remaining-subscription-writes:
       - '1199'
       x-msedge-ref:
-      - 'Ref A: B3BAD7FD309C4333939565D8E0E716FF Ref B: CO6AA3150217025 Ref C: 2024-04-19T21:23:05Z'
+      - 'Ref A: 49439903044B4707978625C9A1EC9601 Ref B: CO6AA3150220011 Ref C: 2024-04-23T18:17:49Z'
     status:
       code: 200
       message: OK
