+++ src/azure-cli/azure/cli/command_modules/compute_recommender/tests/latest/test_compute_recommender.py
@@ -8,8 +8,8 @@
 from azure.cli.testsdk import ScenarioTest
 
 
-class ComputeSkuRecommenderScenario(ScenarioTest):
-    def test_spot_placement_recommender_generate(self):
+class ComputeRecommenderScenario(ScenarioTest):
+    def test_spot_placement_recommender_generate(self):        
         self.kwargs.update({
             'location': 'eastus',
             'subscription_id': self.get_subscription_id(),
@@ -19,6 +19,6 @@ class ComputeSkuRecommenderScenario(ScenarioTest):
             'desired_sizes': '[{"sku": "Standard_D2_v2"}]'
         })
 
-        spot_scores_output = self.cmd('az compute-sku-recommender spot-placement-recommender -l {location} --subscription {subscription_id} --availability-zones {availability_zones} --desired-locations \'{desired_locations}\' --desired-count {desired_count} --desired-sizes \'{desired_sizes}\'').get_output_in_json()
+        spot_scores_output = self.cmd('az compute-recommender spot-placement-recommender -l {location} --subscription {subscription_id} --availability-zones {availability_zones} --desired-locations \'{desired_locations}\' --desired-count {desired_count} --desired-sizes \'{desired_sizes}\'').get_output_in_json()
 
         self.assertTrue(len(spot_scores_output["placementScores"]) > 0, "Spot scores should have at least one item")
