+++ src/azure-cli/service_name.json
@@ -120,7 +120,7 @@
     "URL": "https://learn.microsoft.com/azure/cognitive-services"
   },
   {
-    "Command": "az compute-sku-recommender",
+    "Command": "az compute-recommender",
     "AzureServiceName": "Compute Diagnostic",
     "URL": ""
   },
